Implement faster ToArray
Previously we were relying on IEnumerable<T>.ToArray which was painfully slow for our use. This dramatically sped up the signature generation from ~120 seconds to 3 when running a profile on it. The main bottleneck in the generation is now XXHash itself which is about as fast as you can get anyway